refactor: split Sidebar.jsx into modular component files (#184)
Extract ShareModal, CellSignal/EditableCellSignal, ReadOnlyView, EditView,
PoiNews, ContentDetail, PoiEvents, AssociationsModal, AssociationsTabContent,
and TrailStatus into frontend/src/components/sidebar/, with shared helpers in
sidebar/helpers.js. Sidebar.jsx drops from 3583 to 1217 lines and is now a
container that composes the extracted pieces. Pure code movement — no logic,
JSX, prop, or behavior changes.

Phase C (part 3) of the frontend unification (spec 019).
